requirements.txt
streamlit>=1.20
google-genai>=1.0.0   # official modern GenAI SDK (recommended)
pydub
python-dotenv
python-docx
fpdf
tqdm
